requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
requests
pillow
python-dateutil
